Feat: Dividing code responsibilites(`jwt provider`, `interfaces`)
