--- src/modules/auth/interfaces/auth.interface.ts
@@ -1,4 +1,0 @@
-interface KakaoLogin {
-  refreshToken: Promise<string>;
-  accessToken: Promise<string>;
-}
